+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/AuthenticationRequestParameters.kt
@@ -1,17 +1,25 @@
 package at.asitplus.openid
 
 import at.asitplus.KmmResult.Companion.wrap
-import at.asitplus.signum.indispensable.josef.io.InstantLongSerializer
 import at.asitplus.dif.PresentationDefinition
+import at.asitplus.signum.indispensable.io.ByteArrayBase64Serializer
+import at.asitplus.signum.indispensable.josef.JsonWebToken
+import at.asitplus.signum.indispensable.josef.io.InstantLongSerializer
 import kotlinx.datetime.Instant
+import kotlinx.serialization.KSerializer
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.descriptors.PrimitiveKind
+import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
+import kotlinx.serialization.descriptors.SerialDescriptor
 import kotlinx.serialization.encodeToString
+import kotlinx.serialization.encoding.Decoder
+import kotlinx.serialization.encoding.Encoder
 
 /**
  * Contents of an OIDC Authentication Request.
  *
- * Usually, these parameters are appended to the Authorization Endpoint URL of the OpenId Provider (may be the
+ * Usually, these parameters are appended to the Authorization Endpoint URL of the OpenId Provider (maybe the
  * Wallet App in case of SIOPv2, or the Credential Issuer for OID4VCI).
  */
 @Serializable
@@ -258,6 +266,82 @@ data class AuthenticationRequestParameters(
      */
     @SerialName("code_challenge_method")
     val codeChallengeMethod: String? = null,
+
+    //CSC additionally defines the following parameters
+    //In the following REQUIRED-"credential" stands for
+    // "It SHALL be used only if
+    // [scope] is 'credential' "
+    /**
+     * CSC: Optional
+     * Request a preferred language according to RFC 5646
+     */
+    @SerialName("lang")
+    val lang: String? = null,
+
+    /**
+     * CSC: REQUIRED-"credential"
+     * The identifier associated to the credential to authorize
+     */
+    @SerialName("credentialID")
+    val credentialID: String? = null,
+
+    /**
+     * CSC: Required-"credential"
+     * This parameter contains the symbolic identifier determining the kind of
+     * signature to be created
+     */
+    @SerialName("signatureQualifier")
+    val signatureQualifier: String? = null,
+
+    /**
+     * CSC: Required-"credential"
+     * The number of signatures to authorize
+     */
+    @SerialName("numSignatures")
+    val numSignatures: Int? = null,
+
+    /**
+     * CSC: REQUIRED-"credential"
+     * One or more base64url-encoded hash values to be signed
+     */
+    @SerialName("hashes")
+    @Serializable(HashesSerializer::class)
+    val hashes: List<ByteArray>? = null,
+
+    /**
+     * CSC: REQUIRED-"credential"
+     * String containing the OID of the hash algorithm used to generate the hashes
+     */
+    @SerialName("hashAlgorithmOID")
+    val hashAlgorithmOID: String? = null,
+
+    /**
+     * CSC: OPTIONAL
+     * A free form description of the authorization transaction in the lang language.
+     * The maximum size of the string is 500 characters
+     */
+    @SerialName("description")
+    val description: String? = null,
+
+    /**
+     * CSC: OPTIONAL
+     * To restrict access to the authorization server of a remote service, this specification introduces the
+     * additional account_token parameter to be used when calling the oauth2/authorize endpoint. This
+     * parameter contains a secure token designed to authenticate the authorization request based on an
+     * Account ID that SHALL be uniquely assigned by the signature application to the signing user or to the
+     * user’s application account
+     */
+    @SerialName("account_token")
+    val accountToken: JsonWebToken? = null,
+
+    /**
+     * CSC: OPTIONAL
+     * Arbitrary data from the signature application. It can be used to handle a
+     * transaction identifier or other application-spe cific data that may be useful for
+     * debugging purposes
+     */
+    @SerialName("clientData")
+    val clientData: String? = null,
 ) {
 
     fun serialize() = jsonSerializer.encodeToString(this)
@@ -266,5 +350,24 @@ data class AuthenticationRequestParameters(
         fun deserialize(it: String) = kotlin.runCatching {
             jsonSerializer.decodeFromString<AuthenticationRequestParameters>(it)
         }.wrap()
+    }
+}
+
+/**
+ * CSC: Multiple hash values can be passed as comma separated values,
+ * e.g. oauth2/authorize?hash=dnN3ZX.. .ZmRm,ZjIxM3… Z2Zk,…
+ */
+object HashesSerializer : KSerializer<List<ByteArray>> {
+    override val descriptor: SerialDescriptor =
+        PrimitiveSerialDescriptor("HashesSerializer", PrimitiveKind.STRING)
+
+    override fun deserialize(decoder: Decoder): List<ByteArray> {
+        val listOfHashes = decoder.decodeString().split(",")
+        return listOfHashes.map { at.asitplus.dif.jsonSerializer.decodeFromString(ByteArrayBase64Serializer, it) }
+    }
+
+    override fun serialize(encoder: Encoder, value: List<ByteArray>) {
+        val listOfHashes = value.map { at.asitplus.dif.jsonSerializer.encodeToString(ByteArrayBase64Serializer, it) }
+        encoder.encodeString(listOfHashes.joinToString(","))
     }
 }